# Conflicts:
#	src/main/java/com/app/cfp/controller/SystemController.java
#	src/main/java/com/app/cfp/service/AllocationService.java
#	src/main/java/com/app/cfp/service/MedicalCaseService.java
#	src/main/java/com/app/cfp/service/SystemService.java

+++ src/main/java/com/app/cfp/service/SchedulerService.java
@@ -2,14 +2,12 @@ package com.app.cfp.service;
 
 import com.app.cfp.entity.MedicalCase;
 import com.app.cfp.entity.TempMedicalCase;
-import com.app.cfp.utils.ImageUtility;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.scheduling.annotation.EnableScheduling;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
 import java.util.Random;
 
 @Service
@@ -20,43 +18,46 @@ public class SchedulerService {
 
     private final MedicalCaseService medicalCaseService;
 
+    private final ResidentService residentService;
+
     private final Random random = new Random();
 
     @Value("${numberOfCasesAutomaticallyAssignedPerDay}")
     private int numberOfCasesAutomaticallyAssignedPerDay;
 
     @Autowired
-    public SchedulerService(SystemService systemService, MedicalCaseService medicalCaseService) {
+    public SchedulerService(SystemService systemService, MedicalCaseService medicalCaseService, ResidentService residentService) {
         this.systemService = systemService;
         this.medicalCaseService = medicalCaseService;
+        this.residentService = residentService;
     }
 
-    //TODO @Cronjob at 14:00:00
-    public void assignEducationalCases() {
-        //TODO
-        //if cases are not enough
-        //getAllEducationalCases
-        //allocateEducationalCaseToResident0
-    }
-
-    @Scheduled(cron = "0 0 0 * * *", zone = "Europe/Istanbul")
+    @Scheduled(cron = "0 0 20 * * *", zone = "Europe/Istanbul")
     public void addTempMedicalCases() {
-        List<TempMedicalCase> tempMedicalCaseList = systemService.getAllTempMedicalCases();
-        for (int i = 0; i < numberOfCasesAutomaticallyAssignedPerDay; i++) {
-            int position = random.nextInt(tempMedicalCaseList.size());
-            TempMedicalCase tempMedicalCase = tempMedicalCaseList.get(position);
-            medicalCaseService.addMedicalCase(createMedicalCase(tempMedicalCase.getCFPImage(), tempMedicalCase.getPresumptiveDiagnosis()));
+        System.out.println("Automatically assign medical cases");
+        long numberOfTempMedicalCases = systemService.countAllTempMedicalCases();
+        long numberOfResidents = residentService.countAllResidents();
+        for (int i = 0; i < numberOfCasesAutomaticallyAssignedPerDay * numberOfResidents; i++) {
+            if (numberOfTempMedicalCases == 0L){
+                return;
+            }
+//            int position = random.nextInt(tempMedicalCaseList.size());
+            TempMedicalCase tempMedicalCase = systemService.findRandomTempMedicalCase();
+            medicalCaseService.addMedicalCase(createMedicalCase(tempMedicalCase.getCFPImage(), tempMedicalCase.getPresumptiveDiagnosis(), tempMedicalCase.getCFPImageName()));
             systemService.deleteTempMedicalCase(tempMedicalCase.getId());
-            tempMedicalCaseList.remove(position);
+//            tempMedicalCaseList.remove(position);
         }
     }
 
-    private MedicalCase createMedicalCase(byte[] fileData, String presumtiveDiagnosis) {
+    private MedicalCase createMedicalCase(byte[] fileData, String presumtiveDiagnosis, String CFPImageName) {
         MedicalCase medicalCase = new MedicalCase();
-        medicalCase.setEncodedInfo("$2a$12$MDnofLJT8LrIILyh8SCle.DN9yKFRqQg8W.dqbq6hlkKqE0bL6mKK");
-        medicalCase.setAdditionalInformation("INFORMATII ADITIONALE!");
-        medicalCase.setCFPImage(ImageUtility.compressImage(fileData));
+        String randomName = java.util.UUID.randomUUID() + java.util.UUID.randomUUID().toString();
+
+        medicalCase.setEncodedInfo("$2a$12$" + randomName.substring(0, 52));
+        medicalCase.setAdditionalInformation("No additional information");
+        medicalCase.setCFPImage(fileData);
         medicalCase.setPresumptiveDiagnosis(presumtiveDiagnosis);
+        medicalCase.setCFPImageName(CFPImageName);
 
         return medicalCase;
     }
